added exactPlayoutValue option ( use [-1,1] value instead of {-1,1})
git-svn-id: svn+ssh://zene.sk/home/tomik/svn/akimot/trunk@136 6bcf84ba-f5f3-4bf5-b896-c696636b5bce

+++ uct.h
@@ -431,7 +431,7 @@ class Uct:public Engine
      *         If winner is not known ( no winning criteria reached ), position
      *         is evaluated and biased coin si flipped to estimate the winner
      */
-    int decidePlayoutWinner(const Board*) const;
+    double decidePlayoutWinner(const Board*) const;
 
     /**
      * Filtering steps through Transposition tables.
